requirements.txt
langchain
openai
openpyxl
pandas
toml
pdfminer.six